+++ packages/plugin-packages/mashroom-portal/test/backend/utils/path_utils.test.js
@@ -62,6 +62,29 @@ describe('path_utils', () => {
         expect(getFrontendSiteBasePath(req)).toBe('/mashroom-portal');
     });
 
+    it('resolves the frontend site base path when virtual host frontend base path is root', () => {
+        const req: any = {
+            params: {
+                sitePath: 'web1',
+            },
+            pluginContext: {
+                services: {
+                    vhostPathMapper: {
+                        service: {
+                            getMappingInfo() {
+                                return {
+                                    frontendBasePath: '/',
+                                };
+                            },
+                        },
+                    },
+                },
+            },
+        };
+
+        expect(getFrontendSiteBasePath(req)).toBe('');
+    });
+
     it('resolves the frontend API and resources base path', () => {
         const req: any = {
             params: {