+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## [unreleased]
 
+ * Portal: Fixed mapping a Site to a virtual host when the frontend base path is root
  * Virtual host path mapper: Added a config property to explicitly set the http headers that
    should be considered (default is x-forwarded-host).